requirements.txt
asgiref==3.2.3
attrs==21.2.0
common==0.1.2
Django==3.0.4
django-appconf==1.0.5
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.9.0
django-environ==0.4.5
django-select2==7.7.2
gevent==21.8.0
greenlet==1.1.1
iniconfig==1.1.1
mock==4.0.3
numpy==1.21.4
packaging==21.0
Pillow==7.1.1
pluggy==1.0.0
psycopg2-binary==2.8.4
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytz==2019.3
six==1.16.0
sqlparse==0.3.1
toml==0.10.2
torch==1.10.0
torchvision==0.11.1
typing_extensions==4.0.0
zope.event==4.5.0
zope.interface==5.4.0
